pyproject.toml
[project]
name = "fintech-ledger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "unidecode>=1.4.0",
]
